requirements.txt
# Core Framework
streamlit>=1.35.0

# AI APIs
google-genai>=1.0.0
groq>=1.0.0

# PDF Processing
pdfplumber>=0.11.0
pypdf>=4.0.0

# Image Processing & OCR
Pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.17.0

# Document Processing
python-docx>=1.1.0

# PDF Generation
reportlab>=4.2.0
fpdf2>=2.7.9
